wip : refactor to have new setup
