+++ CMakeLists.txt
@@ -13,6 +13,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED True)
 option(ENABLE_TESTING "Enable Test Builds" ON)
 
 add_subdirectory(utils)
+add_subdirectory(communication)
 
 
 add_executable(${PROJECT_NAME} main.cpp)
@@ -20,6 +21,7 @@ add_executable(${PROJECT_NAME} main.cpp)
 target_link_libraries(${PROJECT_NAME} 
     PRIVATE
     utils
+    communication
 )
