+++ communication/src/uart.cpp
@@ -3,7 +3,7 @@
 #include <errno.h>   // Error integer and strerror() function
 #include <termios.h> // Contains POSIX terminal control definitions
 #include <unistd.h>  // write(), read(), close()
-#include "utils/logger.hpp"
+#include "logger.h"
 
 namespace com {
   uart::uart() : file_desc(-1) {};
