+++ main.cpp
@@ -1,7 +1,7 @@
-#include "utils/logger.h"
-#include "communication/message.h"
-#include "communication/uart.h"
-#include "communication/handler.h"
+#include "logger.h"
+#include "message.h"
+#include "uart.h"
+#include "handler.h"
 #include <memory>
